Conflicts:
	Sources/Internal/Entity/SceneSystem.h
	Sources/Internal/Math/Matrix4.h
	Sources/Internal/Render/Highlevel/RenderBatch.cpp
	Sources/Internal/Render/OGLHelpers.h
	Sources/Internal/Render/OcclusionQuery.cpp
	Sources/Internal/Render/RenderOptions.h

+++ Sources/Internal/Scene3D/Components/LodComponent.cpp
@@ -71,6 +71,7 @@ float32 LodComponent::LodDistance::GetFarDistance() const
 	return sqrtf(farDistanceSq);
 }
 
+
 Component * LodComponent::Clone(Entity * toEntity)
 {
 	LodComponent * newLod = new LodComponent();
@@ -319,7 +320,7 @@ void LodComponent::SetLayerVisibility(int32 layerNum, bool visible)
 {
 	DVASSERT(0 <= layerNum && layerNum < MAX_LOD_LAYERS);
 
-	if(lodLayers.size() > layerNum)
+	if((int32)lodLayers.size() > layerNum)
 	{
 		int32 size = lodLayers[layerNum].nodes.size();
 		for (int32 i = 0; i < size; i++)
